Improve tests compile time
