+++ test/msg_cache_unittest.cpp
@@ -38,6 +38,9 @@
 #include "message_filters/cache.hpp"
 #include "message_filters/message_traits.hpp"
 
+namespace
+{
+
 struct Msg
 {
   std_msgs::msg::Header header;
@@ -481,12 +484,4 @@ TEST(Cache, outOfOrderInsertionSorted)
   EXPECT_EQ(data[1]->data, 2);
   EXPECT_EQ(data[2]->data, 3);
 }
-
-int main(int argc, char ** argv)
-{
-  testing::InitGoogleTest(&argc, argv);
-  rclcpp::init(argc, argv);
-  auto ret = RUN_ALL_TESTS();
-  rclcpp::shutdown();
-  return ret;
-}
+}  // namespace
