+++ test/test_approximate_time_policy.cpp
@@ -38,6 +38,9 @@
 #include "message_filters/sync_policies/approximate_time.hpp"
 #include "message_filters/message_traits.hpp"
 
+namespace
+{
+
 struct Header
 {
   rclcpp::Time stamp;
@@ -50,6 +53,7 @@ struct Msg
 };
 typedef std::shared_ptr<Msg> MsgPtr;
 typedef std::shared_ptr<Msg const> MsgConstPtr;
+}  // namespace
 
 namespace message_filters
 {
@@ -67,6 +71,9 @@ struct TimeStamp<Msg>
 }  // namespace message_filters
 
 typedef std::pair<rclcpp::Time, rclcpp::Time> TimePair;
+
+namespace
+{
 typedef std::pair<rclcpp::Time, unsigned int> TimeAndTopic;
 struct TimeQuad
 {
@@ -531,14 +538,4 @@ TEST(ApproxTimeSync, RateBound) {
   sync_test2.sync_.setInterMessageLowerBound(0, s * 2);
   sync_test2.run();
 }
-
-
-int main(int argc, char ** argv)
-{
-  testing::InitGoogleTest(&argc, argv);
-  rclcpp::init(argc, argv);
-
-  auto ret = RUN_ALL_TESTS();
-  rclcpp::shutdown();
-  return ret;
-}
+}  // namespace
