+++ test/test_chain.cpp
@@ -33,6 +33,9 @@
 #include <rclcpp/rclcpp.hpp>
 #include "message_filters/chain.hpp"
 
+namespace
+{
+
 struct Msg
 {
 };
@@ -169,12 +172,4 @@ TEST(Chain, retrieveBaseClass)
   ASSERT_TRUE(c.getFilter<message_filters::PassThrough<Msg>>(0));
   ASSERT_TRUE(c.getFilter<PTDerived>(0));
 }
-
-int main(int argc, char ** argv)
-{
-  testing::InitGoogleTest(&argc, argv);
-  rclcpp::init(argc, argv);
-  auto ret = RUN_ALL_TESTS();
-  rclcpp::shutdown();
-  return ret;
-}
+}  // namespace
