+++ test/test_input_aligner.cpp
@@ -38,6 +38,9 @@
 #include <rclcpp/rclcpp.hpp>
 #include "message_filters/input_aligner.hpp"
 
+namespace
+{
+
 class InputAligner;
 
 struct Header
@@ -56,6 +59,7 @@ struct Msg2
   Header header;
   int data;
 };
+}  // namespace
 
 namespace message_filters
 {
@@ -112,6 +116,9 @@ public:
   rclcpp::Duration update_rate_;
   std::vector<int> cb_content_;
 };
+
+namespace
+{
 
 TEST_F(InputAlignerTest, init)
 {
@@ -408,3 +415,5 @@ TEST_F(InputAlignerTest, get_queue_status)
   EXPECT_EQ(status_1.msgs_processed, 1);
   EXPECT_EQ(status_1.msgs_dropped, 0);
 }
+
+}  // namespace
