+++ test/time_synchronizer_unittest.cpp
@@ -35,6 +35,9 @@
 #include "message_filters/message_traits.hpp"
 #include <rclcpp/rclcpp.hpp>
 
+namespace
+{
+
 struct Header
 {
   rclcpp::Time stamp;
@@ -47,6 +50,7 @@ struct Msg
 };
 typedef std::shared_ptr<Msg> MsgPtr;
 typedef std::shared_ptr<Msg const> MsgConstPtr;
+}  // namespace
 
 namespace message_filters
 {
@@ -84,6 +88,9 @@ public:
   int32_t count_;
   int32_t drop_count_;
 };
+
+namespace
+{
 
 TEST(TimeSynchronizer, compile2)
 {
@@ -558,12 +565,4 @@ TEST(TimeSynchronizer, connectConstructor)
   pt2.add(m);
   ASSERT_EQ(h.count_, 1);
 }
-
-int main(int argc, char ** argv)
-{
-  testing::InitGoogleTest(&argc, argv);
-  rclcpp::init(argc, argv);
-  auto ret = RUN_ALL_TESTS();
-  rclcpp::shutdown();
-  return ret;
-}
+}  // namespace
